==> notebooks/00-hryniewski-download-datasets.ipynb <==
Ce code permet de changer le répertoire de travail du notebook vers le répertoire parent du répertoire actuel. Cela est nécessaire pour que le notebook puisse accéder aux fichiers et modules situés dans le répertoire du projet.

import os
current_dir = %pwd
project_dir = os.path.dirname(current_dir)
%cd $project_dir

# 00 - Téléchargement de Datasets

Afin de commencer l'exploration et l'analyse des bases de données, téléchargez ces quatre datasets.

Name | Size | Images | Download
-----|------|--------|-----------
Text extraction for OCR| 33.5 Mo | 520 | [Google Drive](https://drive.google.com/file/d/1w0FhoxyHAjFrWJBQ63JiJFPBsxBEUOVO/view?usp=drive_link)
Projet OCR classification | 203.3 Mo | 1 608 | [Google Drive](https://drive.google.com/file/d/1wDa-pXwdUmEpubo8UjGwQCWzZt9PNZbI/view?usp=drive_link)
RVL-CDIP Dataset (5% of original dataset)| 1.8 Go | 20 000 | [Google Drive](https://drive.google.com/file/d/13V-hUMebr5PZjqNcuwxUPlB6u7lEq_gB/view?usp=drive_link)
PRADO | 139.6 Mo | 1 589 | [Google Drive](https://drive.google.com/file/d/1Seii3yeWKoc4f9eUNcgAwDtFleO0yjUm/view?usp=drive_link)

Nous avons extrait 5% de la base de données RVL-CDIB, qui initialement contient 400 000 images, pour économiser du temps et des ressources, tout en maintenant le même ratio de différentes catégories par rapport à la base de données initiale.

### Extraction de données :

Remplacez les chemin des variables `data_01_path`, `data_02_path`, `data_03_path`, `data_04_path`.

# Text extraction for OCR
data_01_path = "path/to/data_01.tar.gz"

# Projet OCR classification
data_02_path = "path/to/data_02.tar.gz"

# RVL-CDIP Dataset
data_03_path = "path/to/data_03.tar.gz"

# PRADO
data_04_path = "path/to/data_04.tar.gz"

from src import extract_tar_gz

# Extraire tout les datastes dans le repertoire data/raw
extract_tar_gz(data_01_path, 'data/raw')
extract_tar_gz(data_02_path, 'data/raw')
extract_tar_gz(data_03_path, 'data/raw')
extract_tar_gz(data_04_path, 'data/raw')
