==> notebooks/06-fernandez-concat-text_process.ipynb <==
# Construction des fichiers pour la définition du pipeline de processing textuel

Ce notebook permet d'assembler les fichiers text_process.csv des datasets 1, 3 et 4 en un fichier qui nous servira pour la définition du pipeline de processing textuel


import os
current_dir = %pwd
project_dir = os.path.dirname(current_dir)
%cd $project_dir

Dans le dossier "data", nous allons créer un dossier "processed" dans lequel nous stockerons les fichiers.

import os
import pandas as pd

if not os.path.exists("data/processed"):
    os.makedirs("data/processed")
    print(f"Folder created: 'data/processed'")
else:
    print(f"Folder already exists: 'data/processed'")

Ce code permet de concaténer les 3 fichiers text_process.csv en ajoutant la colonne 'data' pour avoir le jeu de données d'origine des images.


li = []

for i in ["data_01", "data_03", "data_04"]:
    df = pd.read_csv(f"data/raw/{i}/text_process.csv")
    df['data'] = i
    li.append(df)

frame_words = pd.concat(li, axis=0, ignore_index=True)
frame_words.to_csv("data/processed/words.csv", index=False)

frame_words.head()

Ce code permet de concaténer les 3 fichiers text_process.csv ne gardant que les colonnes 'words' et 'category' et en ajoutant la colonne 'count' qui correspond au nombre de mots dans le text.


li = []
for i in ["data_01", "data_03", "data_04"]:
    df = pd.read_csv(f"data/raw/{i}/text_process.csv",
                     usecols=['words', 'category'])
    df['words'] = df['words'].fillna('')
    df.insert(loc=1, column='count', value=df['words'].apply(
        lambda x: len(x.split())))
    li.append(df)

frame_count = pd.concat(li, axis=0, ignore_index=True)
#frame_count.to_csv("data/processed/count_words.csv", index=False)
frame_count.head()

Sélection des catégories définies:

data = frame_count[frame_count["category"].isin(['passeport', 'national_identity_card', 'email', 'invoice', 'scientific_publication', 'handwritten'])].copy()
data.reset_index(drop=True, inplace=True)
data

## Split la data en train et test.

from sklearn.model_selection import train_test_split

seed = 42
data['words'] = data['words'].fillna('')
target = data['category']
features = data.drop('category', axis=1)

X_train, X_test, y_train, y_test = train_test_split(
    features, target, test_size=0.2, random_state=seed)

df_train = pd.concat((X_train, y_train), axis=1).reset_index(drop=True)
df_test = pd.concat((X_test, y_test), axis=1).reset_index(drop=True)

Création d'un dictionnaire avec les mots les plus communs pour chaque catégorie.

from collections import Counter
from src import save_jsonfile

dic_counter = {}
for cat in df_train['category'].unique():
    df_cat = df_train[df_train['category'] == cat]
    cat_words = ' '.join(df_cat['words'].values)
    cat_count = Counter(cat_words.split(' '))
    most_common = cat_count.most_common()
    most_common = [(word, count) for word, count in most_common if word != ""]
    dic_counter[cat] = most_common

save_jsonfile('data/processed/most_common_words.json', dic_counter, indent=4)

Création de la structure de chaque document, en cherchant la diversité lexicale et la densité des mots-clefs pour chaque catégorie et pour 5, 10, 25 & 50 mots les plus communs.

from src import calculate_lexical_diversity, calculate_keyword_density

df_train['lexical_diversity'] = df_train['words'].apply(calculate_lexical_diversity)
df_test['lexical_diversity'] = df_test['words'].apply(calculate_lexical_diversity)

for cat in df_train['category'].unique():
    for k in [5, 10, 25, 50]:
        words = [w[0] for w in dic_counter[cat][:k]]
        df_train[f"keyword_{cat}_{k}"] = df_train['words'].apply(calculate_keyword_density, keywords=words)
        df_test[f"keyword_{cat}_{k}"] = df_test['words'].apply(calculate_keyword_density, keywords=words)

df_train.to_csv("data/processed/words_structure_train.csv", index=False)
df_test.to_csv("data/processed/words_structure_test.csv", index=False)

df_train.head()

df_test.head()
