==> notebooks/08-fernandez-choice-models-text.ipynb <==
# 08 - Choix des modèles de classification textuel


import os
current_dir = %pwd
project_dir = os.path.dirname(current_dir)
%cd $project_dir

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import StandardScaler, LabelEncoder
from lazypredict.Supervised import LazyClassifier
from joblib import dump

df_train = pd.read_csv('data/processed/words_structure_train.csv')
df_test = pd.read_csv('data/processed/words_structure_test.csv')
seed = 42

df_train['words'] = df_train['words'].fillna('')
df_test['words'] = df_test['words'].fillna('')

X_train = df_train.drop('category', axis=1)
y_train = df_train['category']
X_test = df_test.drop('category', axis=1)
y_test = df_test['category']

# 1. Modèle sur les données words


# Encoding the categories
le = LabelEncoder()
y_train_encoded = le.fit_transform(y_train)
y_test_encoded = le.transform(y_test)

# Save LabelEncoder
dump(le, 'models/target_LabelEncoder.joblib')

# Building the corpus of the train and test data for the TF-IDF vectorizer
X_train_corpus = X_train['words'].tolist()
X_test_corpus = X_test['words'].tolist()

# Learns the vocabulary and the IDF on training data only
tfidf_vectorizer = TfidfVectorizer()
X_train_tfidf = tfidf_vectorizer.fit_transform(X_train_corpus)

# We save the TF-IDF of the train data
dump(tfidf_vectorizer, 'models/TfidfVectorizer.joblib')

# Convert the training data to array because the model doesn't take csr_matrix
X_train_tfidf = X_train_tfidf.toarray()

# Transform and convert the test data
X_test_tfidf = tfidf_vectorizer.transform(X_test_corpus)
X_test_tfidf = X_test_tfidf.toarray()

# Approximatly 6h40
clf_tfidf = LazyClassifier(verbose=0, ignore_warnings=True, custom_metric=None, random_state=seed)
models, predictions = clf_tfidf.fit(
    X_train_tfidf, X_test_tfidf, y_train_encoded, y_test_encoded)

display(models)

dump(clf_tfidf, 'models/lazypredict_tfidf.joblib', compress=5)

# 2. Modèle sur les données de structure


X_train_structure = X_train.drop('words', axis=1)
X_test_structure = X_test.drop('words', axis=1)

# Scaling the features on training data
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train_structure)
X_test_scaled = scaler.transform(X_test_structure)

dump(scaler, 'models/StandardScaler.joblib')

# Approximately 5min

clf_structure = LazyClassifier(
    verbose=0, ignore_warnings=True, custom_metric=None, random_state=seed)

models, predictions = clf_structure.fit(
    X_train_scaled, X_test_scaled, y_train_encoded, y_test_encoded)

display(models)

dump(clf_structure, 'models/lazypredict_structure.joblib')

# 3. Modèle sur l'ensemble word_structure


# Rebuilding the dataframe while keeping the transformations
# Converting to np.array to optimize the performances of the model
X_train_word_structure = pd.concat(
    (pd.DataFrame(X_train_tfidf), pd.DataFrame(X_train_scaled)), axis=1).to_numpy()
X_test_word_structure = pd.concat(
    (pd.DataFrame(X_test_tfidf), pd.DataFrame(X_test_scaled)), axis=1).to_numpy()

from lazypredict.Supervised import CLASSIFIERS

# Take too long to train and don't get great results.
remove = ["CalibratedClassifierCV", "SVC"]
selected_models = []
for model_name, model in CLASSIFIERS:
    if model_name not in remove:
        selected_models.append(model)

# Approximately 2h00
clf_word_structure = LazyClassifier(
    verbose=0, ignore_warnings=True, custom_metric=None, random_state=seed, classifiers=selected_models)
models, predictions = clf_word_structure.fit(
    X_train_word_structure, X_test_word_structure, y_train_encoded, y_test_encoded)
display(models)

dump(clf_word_structure, 'models/lazypredict_words_structure.joblib')
